pyproject.toml
[project]
name = "navigation"
version = "0.1.0"
description = "Multi-agent navigation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "arcade>=3.3.2",
    "manim>=0.19.0",
    "matplotlib>=3.10.3",
    "pandas>=2.3.0",
    "pydantic>=2.11.7",
    "pyyaml>=6.0.2",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
]

[tool.uv.workspace]
members = [
    "navigation",
]
